Crate scylla_macros

Procedural macros for ScyllaDB Rust Driver. Intended for use with the scylla and scylla-cql crates.

Derive Macros

DeserializeRow

Derive macro for the DeserializeRow trait that generates an implementation which deserializes a row with a similar layout to the Rust struct.

DeserializeValue

Derive macro for the DeserializeValue trait that generates an implementation which deserializes a User Defined Type with the same layout as the Rust struct.

SerializeRow

Derive macro for the SerializeRow trait which serializes given Rust structure into bind markers for a CQL statement.

SerializeValue

Derive macro for the SerializeValue trait which serializes given Rust structure as a User Defined Type (UDT).